PulseConsole: first version
This is the initial commit for PulseConsole, a Cocoa based application
to introspect details of a PA server. It does connect to the local
daemon, but currently still fails for remote servers.

It listens to Bonjour services and lists them in a NSPopupBox to select
them.

A second feature is to control all the volumes of all sinks, sources and
streams. This is not done yet.

The PulseAudio logo has been transformed into a .icns.

Lots of refactoring is still missing, and it also won't build on other
machines as the script code is not yet mature.