HAL: Added async message implementation
This is needed because the DeviceStop() callback in our plugin
implementation is likely to be called from the IOProc thread by client
applications. As the PulseAudio threaded mainloop API doesn't like that
at all, we need to trick around and send messages into the thread that
created the plugin in the first place.